New class for c2s activities
