+++ src/Protocol/ActivityPub/Processor.php
@@ -2144,70 +2144,4 @@ class Processor
 
 		return $body;
 	}
-
-	/**
-	 * Create an item array from client to server object data
-	 *
-	 * @param array $object_data
-	 * @param array $application
-	 * @param integer $uid
-	 * @return array
-	 */
-	public static function processC2SContent(array $object_data, array $application, int $uid): array
-	{
-		$owner = User::getOwnerDataById($uid);
-
-		$item = [];
-
-		$item['network']    = Protocol::DFRN;
-		$item['uid']        = $uid;
-		$item['verb']       = Activity::POST;
-		$item['contact-id'] = $owner['id'];
-		$item['author-id']  = $item['owner-id'] = Contact::getPublicIdByUserId($uid);
-		$item['title']      = $object_data['name'];
-		$item['body']       = Markdown::toBBCode($object_data['content']);
-		$item['app']        = $application['name'] ?? 'API';
-
-		if (!empty($object_data['target'][Receiver::TARGET_GLOBAL])) {
-			$item['allow_cid'] = '';
-			$item['allow_gid'] = '';
-			$item['deny_cid']  = '';
-			$item['deny_gid']  = '';
-			$item['private']   = Item::PUBLIC;
-		} elseif (isset($object_data['target'][Receiver::TARGET_GLOBAL])) {
-			$item['allow_cid'] = '';
-			$item['allow_gid'] = '';
-			$item['deny_cid']  = '';
-			$item['deny_gid']  = '';
-			$item['private']   = Item::UNLISTED;
-		} elseif (!empty($object_data['target'][Receiver::TARGET_FOLLOWER])) {
-			$item['allow_cid'] = '';
-			$item['allow_gid'] = '<' . Group::FOLLOWERS . '>';
-			$item['deny_cid']  = '';
-			$item['deny_gid']  = '';
-			$item['private'] = Item::PRIVATE;
-		} else {
-			// @todo Set permissions via the $object_data['target'] array
-			$item['allow_cid'] = '<' . $owner['id'] . '>';
-			$item['allow_gid'] = '';
-			$item['deny_cid']  = '';
-			$item['deny_gid']  = '';
-			$item['private'] = Item::PRIVATE;
-		}
-
-		if (!empty($object_data['summary'])) {
-			$item['body'] = '[abstract=' . Protocol::ACTIVITYPUB . ']' . $object_data['summary'] . "[/abstract]\n" . $item['body'];
-		}
-
-		if ($object_data['reply-to-id']) {
-			$item['thr-parent'] = $object_data['reply-to-id'];
-			$item['gravity'] = Item::GRAVITY_COMMENT;
-		} else {
-			$item['gravity'] = Item::GRAVITY_PARENT;
-		}
-
-		$item = DI::contentItem()->expandTags($item);
-
-		return $item;
-	}
 }
